refactor: extract cluster domain models to models/cluster_models.py
