+++ src/slipbox_mcp/services/__init__.py
@@ -1,6 +1,8 @@
 """Service layer for the Zettelkasten MCP server."""
-from slipbox_mcp.services.cluster_service import (  # noqa: F401
+from slipbox_mcp.models.cluster_models import (  # noqa: F401
     ClusterCandidate as ClusterCandidate,
     ClusterReport as ClusterReport,
+)
+from slipbox_mcp.services.cluster_service import (  # noqa: F401
     ClusterService as ClusterService,
 )
